Improved distanceCheck!
distanceCheck! is now able to work with any number of swarms, not just four. Also, the calculation was made more efficient as the sum operation you showed me was implemented. Waiting to update the rest of requested changes until I can make the testing script work. Once that works I can implement the other changes and track how they effect the quality of the MSPSO algorithm.